PPDU duration limitation + time channel is idle
The PPDU duration has been limited according to the IEEE 802.11. In particular, this affects to the number of packets that are tried to be aggregated for a single PPDU. This closes #110.

In  addition, a new statistic has been added, which indicates the total time spent without transmissions (i.e., the total time in which no one is transmitting). We need to improve this measurement to assess the time each channel is idle.