Re-adapt NAV procedure
Some parts of the new NAV implementation (commit c9cc0d525b676174808c5c5a34bfa927bcece675) were lost due to the merge of master into the spatial_reuse branch. We have now solved this issue in this branch, which will be replicated in master as soon as the implementation of the SR operation has been completed successfully.