First results random scenarios + Improvements in the SR operation
First, we have prepared and ordered the scripts to execute the simulations regarding random scenarios.

In addition, we added few changes in the code, so that the AP can notify changes in the SR configuration to STAs. At the beginning of the simulation (Start() node), the AP notifies the STA about the initial configuration. Changes can be done at that moment. For instance, if we want to randomly decide whether a WLAN implements SR or not, this can be easily done at that point.

Finally, other changes have been done with respect to the preparation of random (large-scale) simulations (like in the App "Input generation").