First approach of the action-banning approach
A centralized action-banning mechanism has been implemented, where a central controller keeps track of the performance achieved by every associated agent and bans actions accordingly. The goal of this approach is to boost the performance of decentralized reinforcement learning approaches for  finding the optimal CCA & Tx Power combination, but keeping complexity low.

We will continue working on this and providing other action-banning criteria.