Refactor channel access + EDCA backoff
The channel access has been refactored in order to grant the necessary flexibility to include other policies than DCF. Now, the nodes' input has changed, as well as the "config_models" file. EDCA is implemented only for computing the backoff based on the different types of ACs, but traffic differentiation is still to be implemented. In addition, for EDCA, we still need to implement the corresponding TXOP durations.