CompC++ (#150)
* [CompC++] Checkin source code from sense3.1

COST/CompC++ is not available on the official website anymore. Extract
the compc++ source code from sense3.1
(http://ita.cs.rpi.edu/download/sense-3.1.tar.gz) and check in the
files.

* [CompC++] Add LICENSE file for compc++

Extract Copyright and Disclaimer Notice from the sense website and add
it as LICENSE file.

* [CompC++] Add brief README

Add a brief README that clarifies the purpose of adding compc++.

* [CompC++] Add CompC++ section to gitignore

Ignore binary cxx and generated lexer.cpp files.

* [CompC++] Adjust Makefile

Adjust the compc++ Makefile. The Makefile included in sense includes mk
files from the parent directory. Extract the variables defined in these
files and add them to the Makefile. Adjust the targets to the included
variables.

* [make] Add Makefile to build komondor_main

Create a Makefile with targets for compiling komondor. Add targets to
invoke the CompC++ build as well.

* [Docu] Adjust the documentation

Add section that explains the build process with make.

* [CompC++] Rename directory and add docu

Rename compc++ directory to compcpp since it is bad practice to use
special characters in file names.

Also add a small documentation how to build CompC++ on it's own.